remove category and move query to seperate file
